Alterar paleta de cores via PriceWhisperAppBar
