+++ app/src/main/java/br/com/pricewhisper/MainActivity.kt
@@ -9,17 +9,9 @@ import androidx.activity.viewModels
 import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
@@ -34,6 +26,7 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.navigation
 import androidx.navigation.compose.rememberNavController
 import br.com.pricewhisper.models.Product
+import br.com.pricewhisper.ui.components.PriceWhisperAppBar
 import br.com.pricewhisper.ui.screens.HomeScreen
 import br.com.pricewhisper.ui.screens.products.ProductDetailsScreen
 import br.com.pricewhisper.ui.screens.products.ProductFormScreen
@@ -63,7 +56,7 @@ class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        
+
         lifecycleScope.launch {
             themeViewModel.value.currentPrefs.value = getPreferences(Context.MODE_PRIVATE)
             themeViewModel.value.isDarkModeOn.value = themeViewModel.value.getCurrentPalette()
@@ -81,36 +74,6 @@ class MainActivity : ComponentActivity() {
         }
     }
 
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
-    fun PriceWhisperAppBar(
-        modifier: Modifier = Modifier,
-        currentScreen: PriceWhisperScreen,
-        canNavigateBack: Boolean = false,
-        navigateUp: () -> Unit = {}
-    ) {
-        TopAppBar(
-            title = { Text(text = getString(currentScreen.title)) },
-            colors = TopAppBarDefaults.mediumTopAppBarColors(
-                containerColor = MaterialTheme.colorScheme.primary,
-                titleContentColor = MaterialTheme.colorScheme.onPrimary
-            ),
-            navigationIcon = {
-                if (canNavigateBack) {
-                    IconButton(onClick = navigateUp) {
-                        Icon(
-                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                            contentDescription = stringResource(R.string.top_app_bar_navigation_icon_description),
-                            tint = MaterialTheme.colorScheme.onPrimary,
-                            modifier = Modifier.padding(end = 16.dp)
-                        )
-                    }
-                }
-            },
-            modifier = modifier
-        )
-    }
-
     @Composable
     fun PriceWhisperApp(
         viewModel: Lazy<ProductViewModel>,
@@ -126,7 +89,8 @@ class MainActivity : ComponentActivity() {
                 PriceWhisperAppBar(
                     currentScreen = currentScreen,
                     canNavigateBack = navController.previousBackStackEntry != null,
-                    navigateUp = { navController.navigateUp() }
+                    navigateUp = { navController.navigateUp() },
+                    onSwitchTheme = { themeViewModel.value.switchPalette() }
                 )
             }
         ) { innerPadding ->
@@ -143,7 +107,7 @@ class MainActivity : ComponentActivity() {
                     composable(PriceWhisperScreen.HomeScreen.name) {
                         HomeScreen(
                             onClickProductsActionButton = {
-                                themeViewModel.value.switchPalette()
+                                navController.navigate(PriceWhisperScreen.ProductListScreen.name)
                             },
                             onClickProfileActionButton = {},
                             onClickSettingsActionButton = {}
