+++ app/src/main/res/values/strings.xml
@@ -21,4 +21,5 @@
     <string name="productlist_item_icon_description_delete">Delete Product</string>
     <string name="top_app_bar_navigation_icon_description">Go Back</string>
     <string name="loading_text">Loading…</string>
+    <string name="switch_theme_description">Change App Theme</string>
 </resources>
